backend/requirements.txt
flask 
flask-cors 
elasticsearch 
python-dotenv